backend/requirements.txt
fastapi
uvicorn
psycopg
psycopg-binary
langgraph
langchain-groq
pydantic
pydantic-settings
python-dotenv
